move bearer token code to own file
