+++ src/eduid/scimapi/routers/login.py
@@ -3,7 +3,7 @@ import datetime
 from fastapi import Response
 from jwcrypto import jwt
 
-from eduid.common.models.webauthn import AuthSource
+from eduid.common.models.bearer_token import AuthSource
 from eduid.scimapi.api_router import APIRouter
 from eduid.scimapi.context_request import ContextRequest
 from eduid.scimapi.exceptions import ErrorDetail, NotFound, Unauthorized
