+++ src/eduid/scimapi/tests/test_authn.py
@@ -11,10 +11,9 @@ from jwcrypto import jwt
 
 from eduid.common.config.base import DataOwner, DataOwnerName, ScopeName
 from eduid.common.config.parsers import load_config
+from eduid.common.models.bearer_token import AuthnBearerToken, AuthSource, RequestedAccessDenied, SudoAccess
 from eduid.common.models.scim_base import SCIMSchema
-from eduid.common.models.webauthn import AuthSource, RequestedAccessDenied, SudoAccess
 from eduid.scimapi.config import ScimApiConfig
-from eduid.scimapi.middleware import AuthnBearerToken
 from eduid.scimapi.testing import BaseDBTestCase
 from eduid.scimapi.tests.test_scimuser import ScimApiTestUserResourceBase
 from eduid.userdb.scimapi import ScimApiProfile
